backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.2
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
aiofiles==23.2.1
python-dateutil==2.8.2
pydantic-settings==2.1.0
alembic==1.13.1
